finsihed dao and service checks. minor changes

+++ src/main/java/com/epam/app/dao/CrudRepository.java
@@ -20,7 +20,7 @@ public interface CrudRepository<T> {
      * @return the entity if the needed
      * entity was found, otherwise null.
      */
-    T find(int id);
+    T find(Long id);
 
     /**
      * Updates an entity.
